Ajustes types e juntar exercise e exercisedetailsnav
