+++ src/app/components/screens/Train/Models/ModelNav.tsx
@@ -1,8 +1,11 @@
 import { createStackNavigator } from "@react-navigation/stack"
 import React from "react"
-import { Exercise, TrainingModel } from "../../../../lib/types/train"
+import {
+  Exercise as ExerciseType,
+  TrainingModel
+} from "../../../../lib/types/train"
+import { Exercise } from "../Exercises/ExerciseDetails/Exercise"
 import ExerciseSelector from "../Exercises/ExerciseSelector"
-import { Exercise as ExerciseScreen } from "../Exercises/Exercise"
 import Model, { modelModes } from "./Model"
 import ModelList from "./ModelList"
 import Session from "./ModelSession"
@@ -11,8 +14,8 @@ export type RootStackParamListModelNav = {
   ModelList: undefined;
   Session: { model: TrainingModel; id?: string };
   Model: { model?: TrainingModel; id?: string; mode: modelModes };
-  ExerciseSelector: { onSubmit: (selectedExercises: Exercise[]) => void };
-  Exercise: { exercise: Exercise };
+  ExerciseSelector: { onSubmit: (selectedExercises: ExerciseType[]) => void };
+  Exercise: { exercise: ExerciseType };
 };
 
 export default function ExerciseNav() {
@@ -27,7 +30,7 @@ export default function ExerciseNav() {
       <Stack.Screen name="Model" component={Model} />
       <Stack.Screen name="Session" component={Session} />
       <Stack.Screen name="ExerciseSelector" component={ExerciseSelector} />
-      <Stack.Screen name="Exercise" component={ExerciseScreen} />
+      <Stack.Screen name="Exercise" component={Exercise} />
     </Stack.Navigator>
   )
 }
